+++ vespajlib/src/test/java/com/yahoo/tensor/MappedTensorTestCase.java
@@ -1,6 +1,7 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.tensor;
 
+import com.google.common.collect.Sets;
 import org.junit.Test;
 
 import java.util.Set;
@@ -17,38 +18,23 @@ import static org.junit.Assert.fail;
 public class MappedTensorTestCase {
 
     @Test
-    public void testStringForm() {
-        assertEquals("{}", Tensor.from("{}").toString());
-        assertEquals("{{d1:l1,d2:l1}:5.0,{d1:l1,d2:l2}:6.0}", Tensor.from("{ {d1:l1,d2:l1}: 5,   {d2:l2, d1:l1}:6.0} ").toString());
-        assertEquals("{{d1:l1,d2:l1}:-5.3,{d1:l1,d2:l2}:0.0}", Tensor.from("{ {d1:l1,d2:l1}:-5.3, {d2:l2, d1:l1}:0}").toString());
+    public void testOneDimensionalBuilding() {
+        TensorType type = new TensorType.Builder().mapped("x").build();
+        Tensor tensor = Tensor.Builder.of(type).
+                cell().label("x", "0").value(1).
+                cell().label("x", "1").value(2).build();
+        assertEquals(Sets.newHashSet("x"), tensor.type().dimensionNames());
+        assertEquals("{{x:0}:1.0,{x:1}:2.0}", tensor.toString());
     }
 
     @Test
-    public void testParseError() {
-        try {
-            Tensor.from("--");
-            fail("Expected parse error");
-        }
-        catch (IllegalArgumentException expected) {
-            assertEquals("Excepted a number or a string starting by { or tensor(, got '--'", expected.getMessage());
-        }
-    }
-
-    @Test
-    public void testDimensions() {
-        Set<String> dimensions1 = Tensor.from("{} ").type().dimensionNames();
-        assertEquals(0, dimensions1.size());
-
-        Set<String> dimensions2 = Tensor.from("{ {d1:l1, d2:l2}:5, {d1:l2, d2:l2}:6.0} ").type().dimensionNames();
-        assertEquals(2, dimensions2.size());
-        assertTrue(dimensions2.contains("d1"));
-        assertTrue(dimensions2.contains("d2"));
-
-        Set<String> dimensions3 = Tensor.from("{ {d1:l1, d2:l1, d3:l1}:5, {d1:l1, d2:l2, d3:l1}:6.0} ").type().dimensionNames();
-        assertEquals(3, dimensions3.size());
-        assertTrue(dimensions3.contains("d1"));
-        assertTrue(dimensions3.contains("d2"));
-        assertTrue(dimensions3.contains("d3"));
+    public void testTwoDimensionalBuilding() {
+        TensorType type = new TensorType.Builder().mapped("x").mapped("y").build();
+        Tensor tensor = Tensor.Builder.of(type).
+                cell().label("x", "0").label("y", "0").value(1).
+                cell().label("x", "1").label("y", "0").value(2).build();
+        assertEquals(Sets.newHashSet("x", "y"), tensor.type().dimensionNames());
+        assertEquals("{{x:0,y:0}:1.0,{x:1,y:0}:2.0}", tensor.toString());
     }
 
 }
